Point Ruby BigDecimal to a page with info on
The current link to Ruby BigDecimal goes to a page listing the library classes/methods without any explanation as to the BigDecimal class or how to use it. This new link takes you directly to the BigDecimal page, which explains the concepts, how to use it, and everything else that is listed on the original page.